Terceiro endpoint: ESTATISTICA. Criado um novo model para gravar as estatisticas, mas a logica dos 60 segundos ainda vao ser implementadas.

+++ src/main/java/com/jan1ooo/apirest/dto/TransacaoDTO.java
@@ -3,11 +3,10 @@ package com.jan1ooo.apirest.dto;
 import jakarta.validation.constraints.Past;
 import jakarta.validation.constraints.Positive;
 
-import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
 public record TransacaoDTO(@Positive
-                           BigDecimal valor,
+                           Double valor,
                            @Past
                            LocalDateTime dataHora) {
 }
